requirements.txt
streamlit
langchain
langchain-community
ollama
pillow